TST/REF: directories for test_period, test_datetimes, test_timedeltas (#39472)
